fix:修復地圖失效問題、拆分元件
