+++ src/pages/StationInfo.jsx
@@ -23,8 +23,6 @@ import {
   TileLayer,
   Marker,
   Popup,
-  useMap,
-  useMapEvent,
 } from "react-leaflet";
 import "leaflet/dist/leaflet.css";
 import L from "leaflet";
